Added Course Table

+++ src/main/java/pl/agh/edu/io/Class/ClassSessionNotFoundException.java
@@ -1,7 +1,7 @@
 package pl.agh.edu.io.Class;
 
 public class ClassSessionNotFoundException extends RuntimeException {
-    public ClassSessionNotFoundException(int id) {
+    public ClassSessionNotFoundException(long id) {
         super("Class with id: " + id + " not found");
     }
 }
