+++ src/main/java/pl/agh/edu/io/Class/ClassSessionRepository.java
@@ -2,5 +2,5 @@ package pl.agh.edu.io.Class;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClassSessionRepository extends JpaRepository<ClassSession, Integer> {
+public interface ClassSessionRepository extends JpaRepository<ClassSession, Long> {
 }
