+++ src/main/java/pl/agh/edu/io/Classroom/ClassroomDto.java
@@ -2,6 +2,6 @@ package pl.agh.edu.io.Classroom;
 
 import java.util.List;
 
-public record ClassroomDto(int id, String building, int number, int floor, int capacity,
+public record ClassroomDto(long id, String building, int number, int floor, int capacity,
                            boolean hasComputers, List<String> softwareNames) {
 }
