+++ src/main/java/pl/agh/edu/io/Classroom/ClassroomInUseException.java
@@ -1,7 +1,7 @@
 package pl.agh.edu.io.Classroom;
 
 public class ClassroomInUseException extends RuntimeException {
-    public ClassroomInUseException(int id) {
+    public ClassroomInUseException(long id) {
         super("Clasroom with id: " + id + " has scheduled classes");
     }
 }
