+++ src/main/java/pl/agh/edu/io/Classroom/ClassroomNotFoundException.java
@@ -1,7 +1,7 @@
 package pl.agh.edu.io.Classroom;
 
 public class ClassroomNotFoundException extends RuntimeException {
-    public ClassroomNotFoundException(int id) {
+    public ClassroomNotFoundException(long id) {
         super("Classroom with id: " + id + " not found");
     }
 }
