+++ src/main/java/pl/agh/edu/io/Classroom/ClassroomRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ClassroomRepository extends JpaRepository<Classroom, Integer> {
+public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
     Optional<Classroom> findByBuildingAndNumber(String building, int number);
 }
