+++ src/main/java/pl/agh/edu/io/User/UserDto.java
@@ -1,4 +1,4 @@
 package pl.agh.edu.io.User;
 
-public record UserDto(int id, String firstName, String lastName, String email, String phoneNumber, UserRole role) {
+public record UserDto(long id, String firstName, String lastName, String email, String phoneNumber, UserRole role) {
 }
