+++ src/main/java/pl/agh/edu/io/User/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package pl.agh.edu.io.User;
 
 public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(int id) {
+    public UserNotFoundException(long id) {
         super("User with id: " + id + " not found");
     }
 }
